+++ .gitlab-ci.yml
@@ -135,9 +135,9 @@ release_binary_windows:
     - .\build-win-ci.bat $CI_COMMIT_TAG
     - python ../utils/github.py upload ./release/specterd-$CI_COMMIT_TAG-win64.zip
     - cd release
-    - sha256sum specterd-$CI_COMMIT_TAG-win64.zip > SHA256SUMS-win
-    - type SHA256SUMS-win
-    - echo %GPG_PASSPHRASE% | gpg --detach-sign --armor --no-tty --batch --yes --passphrase-fd 0  --pinentry-mode loopback SHA256SUMS-win
+    - python3 -m utils.release-helper sha256sums specterd-%CI_COMMIT_TAG%-win64.zip > SHA256SUMS-windows
+    - type SHA256SUMS-windows
+    - echo %GPG_PASSPHRASE% | gpg --detach-sign --armor --no-tty --batch --yes --passphrase-fd 0  --pinentry-mode loopback SHA256SUMS-windows
     - dir
   artifacts:
     when: always
@@ -222,12 +222,12 @@ release_signatures:
     - ./utils/create-gitlab-cli-cfg.sh
     - gpg --import --no-tty --batch --yes /credentials/private.key
   script:
-    - python3 -m utils.download-gitlab-artifacts download          # downloads the job-artifacts from gitlab
-    - python3 -m utils.download-gitlab-artifacts downloadgithub    # downloads additional artifacts from github (if not there and is they have SHA256SUMS-something)
-    - python3 -m utils.download-gitlab-artifacts checkhashes       # checks all SHA256SUM* files
-    - python3 -m utils.download-gitlab-artifacts checksig          # checks the signatures of all SHA256SUMM*.asc files
-    - python3 -m utils.download-gitlab-artifacts create            # creates a SHA256SUM.txt
-    - python3 -m utils.download-gitlab-artifacts upload            # uploads it to github
+    - python3 -m utils.release-helper download          # downloads the job-artifacts from gitlab
+    - python3 -m utils.release-helper downloadgithub    # downloads additional artifacts from github (if not there and is they have SHA256SUMS-something)
+    - python3 -m utils.release-helper checkhashes       # checks all SHA256SUM* files
+    - python3 -m utils.release-helper checksig          # checks the signatures of all SHA256SUMM*.asc files
+    - python3 -m utils.release-helper create            # creates a SHA256SUM.txt
+    - python3 -m utils.release-helper upload            # uploads it to github
     - ./utils/sign_artifact.sh --artifact ./signing_dir/SHA256SUMS
     - python3 ./utils/github.py upload ./signing_dir/SHA256SUMS.asc
